Add AwfulURL class to abstract url parsing.

+++ application/src/com/ferg/awfulapp/AwfulFragment.java
@@ -141,9 +141,9 @@ public abstract class AwfulFragment extends SherlockFragment implements AwfulUpd
 		return (AwfulActivity) getActivity();
 	}
 	
-	protected void displayForum(int forumId, int page){
+	protected void displayForum(long forumId, long page){
 		if(getAwfulActivity() != null){
-			getAwfulActivity().displayForum(forumId, page);
+			getAwfulActivity().displayForum((int)forumId, (int)page);
 		}
 	}
